backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
openai
pydantic
httpx
websockets